backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
mysqlclient==2.0.3
sqlparse==0.4.4
typing_extensions==4.8.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
markdown==3.3.4
django-filter==23.5